Change naming and meta layouts
Want to include the CodePen page for active projects. Keeping the CodePen, repo,
and live web page in sync is going to be a bit of a pain, but it's only for a
while. The bulk of this is probably going to be deleted, or at least moth-balled
at the end of the course.

Minor adjustments to try and keep some semblance of uniformity between the
assorted projects, exercises and assignments within the course. Diff can show
the changes if it of interest later.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>